@@ -71,7 +71,7 @@ dependencies:
   # Code Generation
   flutter_localizations:
     sdk: flutter
-  intl: 0.18.0
+  intl: ^0.18.0
   freezed: ^2.3.2
   acter_flutter_sdk:
     path: packages/rust_sdk
@@ -97,7 +97,7 @@ dependencies:
   file_picker: ^5.2.5
   filesystem_picker: ^3.1.0
   open_app_file: ^4.0.1
-  emoji_picker_flutter: ^1.5.2
+  emoji_picker_flutter: ^1.6.1
   connectivity_plus: ^4.0.1
   mime: ^1.0.2
   permission_handler: ^10.2.0
@@ -114,12 +114,15 @@ dependencies:
   flutter_math_fork: ^0.7.1
   shared_preferences: ^2.2.0
 
-  # Testing
-  mockito: ^5.0.0
   image_picker_for_web: ^2.1.11
   flutter_onboarding_slider: ^1.0.9
   riverpod_infinite_scroll: ^1.0.9
   infinite_scroll_pagination: ^3.2.0
+  awesome_notifications:
+    # path: "../../awesome_notifications"
+    git:
+      url: https://github.com/acterglobal/awesome_notifications
+      ref: acter-main
 
 dev_dependencies:
   flutter_test:
@@ -129,6 +132,7 @@ dev_dependencies:
   icons_launcher: ^2.0.7
   flutter_lints: ^2.0.2
   build_runner: ^2.3.3
+  mockito: ^5.0.0
   # The current build runner version is breaking when generating due to the issue 
   # https://github.com/dart-lang/build/issues/3519, the most probable fix is to upgrade
   # but that requires updated flutter SDK, temporarily we can add build_runner_core
@@ -138,7 +142,7 @@ dev_dependencies:
     # path: "../../flutter_gherkin"
     git:
       url: https://github.com/acterglobal/flutter_gherkin
-      branch: main
+      ref: main
   gherkin: ^3.1.0
 
 icons_launcher:
