@@ -5,22 +5,29 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter/foundation.dart';
 
 class InvitationListNotifier extends Notifier<List<Invitation>> {
-  late Stream<FfiListInvitation> _listener;
+  Stream<FfiListInvitation>? _listener;
   // ignore: unused_field
-  late StreamSubscription<void> _poller;
+  StreamSubscription<void>? _poller;
 
   @override
   List<Invitation> build() {
-    final client = ref.watch(clientProvider)!;
+    final client = ref.watch(clientProvider);
+    if (client == null) {
+      _listener = null;
+      _poller = null;
+      return [];
+    }
     _listener = client.invitationsRx();
-    _poller = _listener.listen((ev) {
-      final asList = ev.toList();
-      debugPrint(
-        ' --- - - ----------------- new invitations received ${asList.length}',
-      );
-      state = asList;
-    });
-    ref.onDispose(() => _poller.cancel());
+    if (_listener != null) {
+      _poller = _listener!.listen((ev) {
+        final asList = ev.toList();
+        debugPrint(
+          ' --- - - ----------------- new invitations received ${asList.length}',
+        );
+        state = asList;
+      });
+      ref.onDispose(() => _poller != null ? _poller!.cancel() : null);
+    }
     return [];
   }
 }
