@@ -68,12 +68,12 @@ impl Convo {
             for event in events {
                 // show only message event as latest message in chat room list
                 // skip the state event
-                if let Ok(AnySyncTimelineEvent::MessageLike(m)) = event.event.deserialize() {
-                    if let Some(msg) = sync_event_to_message(event.clone(), room.clone()) {
-                        self.set_latest_message(msg);
-                        return;
-                    }
+                // if let Ok(AnySyncTimelineEvent::MessageLike(m)) = event.event.deserialize() {
+                if let Some(msg) = sync_event_to_message(&event.event, room.room_id().to_owned()) {
+                    self.set_latest_message(msg);
+                    return;
                 }
+                // }
             }
         }
     }
@@ -246,7 +246,8 @@ impl ConvoController {
                 let ev = raw_event
                     .deserialize_as::<OriginalSyncRoomEncryptedEvent>()
                     .unwrap();
-                let msg = RoomMessage::room_encrypted_from_sync_event(ev, room);
+                let msg =
+                    RoomMessage::room_encrypted_from_sync_event(ev, room.room_id().to_owned());
                 convo.set_latest_message(msg.clone());
 
                 if let Some(idx) = convos.iter().position(|x| x.room_id() == room_id) {
@@ -273,9 +274,14 @@ impl ConvoController {
         if let SdkRoom::Joined(joined) = room {
             let mut convos = self.convos.lock_mut();
             let room_id = room.room_id();
+            let sent_by_me = if let Some(user_id) = room.client().user_id() {
+                ev.sender == user_id
+            } else {
+                false
+            };
 
             let mut convo = Convo::new(Room { room: room.clone() });
-            let msg = RoomMessage::room_message_from_sync_event(ev, room, true);
+            let msg = RoomMessage::room_message_from_sync_event(ev, room_id.to_owned(), sent_by_me);
             convo.set_latest_message(msg.clone());
 
             if let Some(idx) = convos.iter().position(|x| x.room_id() == room_id) {
@@ -305,7 +311,7 @@ impl ConvoController {
                 let room_id = room.room_id();
 
                 let mut convo = Convo::new(Room { room: room.clone() });
-                let msg = RoomMessage::room_member_from_sync_event(ev, room);
+                let msg = RoomMessage::room_member_from_sync_event(ev, room_id.to_owned());
                 convo.set_latest_message(msg.clone());
 
                 if let Some(idx) = convos.iter().position(|x| x.room_id() == room_id) {
@@ -361,7 +367,7 @@ impl ConvoController {
             let room_id = room.room_id();
 
             let mut convo = Convo::new(Room { room: room.clone() });
-            let msg = RoomMessage::room_redaction_from_sync_event(ev, room);
+            let msg = RoomMessage::room_redaction_from_sync_event(ev, room_id.to_owned());
             convo.set_latest_message(msg.clone());
 
             if let Some(idx) = convos.iter().position(|x| x.room_id() == room_id) {
